removed +200 warnings on xcode. didnt change: sqlite and crc32. careful, int32 <> ptr!

+++ TotalCrossVM/src/jpeg/jdatadst.c
@@ -106,7 +106,7 @@ METHODDEF(void)
 term_destination (j_compress_ptr cinfo)
 {
   my_dest_ptr dest = (my_dest_ptr) cinfo->dest;
-  size_t datacount = OUTPUT_BUF_SIZE - dest->pub.free_in_buffer;
+  int datacount = (int)(OUTPUT_BUF_SIZE - dest->pub.free_in_buffer);
 
   /* Write any data remaining in the buffer */
   if (datacount > 0) {
