+++ TotalCrossVM/src/util/posix/utils_c.h
@@ -45,7 +45,7 @@ static int32 privateGetFreeMemory(bool maxblock)
     /* Stats in bytes */
     //flsobral@tc126: kept code to find out the used memory for further reference.
 //    natural_t mem_used = (vm_stat.active_count + vm_stat.inactive_count + vm_stat.wire_count) * pagesize;
-    natural_t mem_free = vm_stat.free_count * pagesize;
+    natural_t mem_free = (int)(vm_stat.free_count * pagesize);
 //    natural_t mem_total = mem_used + mem_free;
     
     return mem_free;
@@ -113,7 +113,7 @@ static int32 privateGetTimeStamp()
 {
    struct timeval now;
    gettimeofday(&now, NULL);
-   return now.tv_sec * 1000 + now.tv_usec / 1000;
+   return (int32)(now.tv_sec * 1000 + now.tv_usec / 1000);
 }
 
 static bool pfileIsDir(TCHARP dir, TCHARP file)
